Change RuleBreakDataV1::property_count from usize to u8
`usize` has different sizes on 32 bit and 64 bit system. Use `u8` so that it has
consistent size across all platforms.

Generate PROPERTY_COUNT as `u8` directly in `build.rs`.

+++ experimental/segmenter/src/provider.rs
@@ -110,7 +110,7 @@ pub struct RuleBreakDataV1<'data> {
     pub break_state_table: RuleBreakStateTable<'data>,
 
     /// Number of properties; should be the square root of the length of [`Self::break_state_table`].
-    pub property_count: usize,
+    pub property_count: u8,
 
     pub last_codepoint_property: i8,
     pub sot_property: u8,
